Moved integration tests to /tests
