+++ src/pmcfg/from_str.rs
@@ -160,108 +160,6 @@ fn parse_composition<T>(input: &[u8]) -> IResult<&[u8], Vec<Vec<VarT<T>>>>
     parse_vec(input, parse_projection, "[", "]", ",")
 }
 
-// TODO move this to integration tests
-#[test]
-fn test_from_str_pmcfg() {
-    use log_domain::LogDomain;
-    use tree_stack_automaton::TreeStackAutomaton;
-    use recognisable::Recognisable;
-
-    let c0: Composition<String> = Composition {
-        composition: vec![vec![VarT::Var(0, 0), VarT::Var(1, 0), VarT::Var(0, 1), VarT::Var(1, 1)]],
-    };
-
-    let c1: Composition<String> = Composition { composition: vec![vec![], vec![]] };
-
-    let c2 = Composition {
-        composition: vec![vec![VarT::T("a".to_string()),
-                               VarT::Var(0, 0)],
-                          vec![VarT::T("c".to_string()),
-                               VarT::Var(0, 1)]],
-    };
-
-    let c3 = Composition {
-        composition: vec![vec![VarT::T("b".to_string()),
-                               VarT::Var(0, 0)],
-                          vec![VarT::T("d".to_string()),
-                               VarT::Var(0, 1)]],
-    };
-
-    let r0: PMCFGRule<String, String, LogDomain<f64>> = PMCFGRule {
-        head: "S".to_string(),
-        tail: vec!["A".to_string(), "B".to_string()],
-        composition: c0.clone(),
-        weight: LogDomain::new(1.0).unwrap(),
-    };
-
-    let r1: PMCFGRule<String, String, LogDomain<f64>> = PMCFGRule {
-        head: "A".to_string(),
-        tail: Vec::new(),
-        composition: c1.clone(),
-        weight: LogDomain::new(0.6).unwrap(),
-    };
-
-    let r2: PMCFGRule<String, String, LogDomain<f64>> = PMCFGRule {
-        head: "A".to_string(),
-        tail: vec!["A".to_string()],
-        composition: c2.clone(),
-        weight: LogDomain::new(0.4).unwrap(),
-    };
-
-    let r3: PMCFGRule<String, String, LogDomain<f64>> = PMCFGRule {
-        head: "B".to_string(),
-        tail: Vec::new(),
-        composition: c1.clone(),
-        weight: LogDomain::new(0.7).unwrap(),
-    };
-
-    let r4: PMCFGRule<String, String, LogDomain<f64>> = PMCFGRule {
-        head: "B".to_string(),
-        tail: vec!["B".to_string()],
-        composition: c3.clone(),
-        weight: LogDomain::new(0.3).unwrap(),
-    };
-
-    let r0_string = "\"S\" → [[Var 0 0, Var 1 0, Var 0 1, Var 1 1]] (\"A\", B)";
-    let r1_string = "A → [[],[]] ()  # 0.6 % this is a comment";
-    let r2_string = "A → [[T a, Var 0 0],[T c, Var 0 1]] (A)  # 0.4";
-    let r3_string = "B → [[],[]] ()  # 0.7";
-    let r4_string = "B → [[T b, Var 0 0],[T d, Var 0 1]] (B)  # 0.3";
-
-    assert_eq!(Ok(r0.clone()),
-               r0_string.parse::<PMCFGRule<String, String, LogDomain<f64>>>());
-    assert_eq!(Ok(r1.clone()),
-               r1_string.parse::<PMCFGRule<String, String, LogDomain<f64>>>());
-    assert_eq!(Ok(r2.clone()),
-               r2_string.parse::<PMCFGRule<String, String, LogDomain<f64>>>());
-    assert_eq!(Ok(r3.clone()),
-               r3_string.parse::<PMCFGRule<String, String, LogDomain<f64>>>());
-    assert_eq!(Ok(r4.clone()),
-               r4_string.parse::<PMCFGRule<String, String, LogDomain<f64>>>());
-
-    let g: PMCFG<String, String, LogDomain<f64>> = PMCFG {
-        initial: vec!["S".to_string()],
-        rules: vec![r0.clone(), r1.clone(), r2.clone(), r3.clone(), r4.clone()],
-    };
-
-    let mut g_string = String::from("initial: [S]\n\n");
-    g_string.push_str(r0_string.clone());
-    g_string.push_str("\n");
-    g_string.push_str(r1_string.clone());
-    g_string.push_str("\n");
-    g_string.push_str(r2_string.clone());
-    g_string.push_str("\n");
-    g_string.push_str(r3_string.clone());
-    g_string.push_str("\n");
-    g_string.push_str(r4_string.clone());
-
-    assert_eq!(Ok(g.clone()), g_string.parse());
-
-    let a = TreeStackAutomaton::from(g);
-
-    assert_ne!(None, a.recognise(vec!["a".to_string(), "b".to_string(), "c".to_string(), "d".to_string()]).next());
-}
-
 #[cfg(test)]
 pub mod tests {
     use super::*;
